# BoundedArray for Hermite and quadratic equations