--- a/numerics/root_finders.hpp
+++ b/numerics/root_finders.hpp
@@ -3,12 +3,14 @@
 
 #include <vector>
 
+#include "base/array.hpp"
 #include "quantities/named_quantities.hpp"
 
 namespace principia {
 namespace numerics {
 namespace internal_root_finders {
 
+using base::BoundedArray;
 using quantities::Derivative;
 
 // Approximates a root of |f| between |lower_bound| and |upper_bound| by
@@ -24,9 +26,8 @@ Argument Bisect(Function f,
 // Returns the solutions of the quadratic equation:
 //   a2 * (x - origin)^2 + a1 * (x - origin) + a0 == 0
 // The result may have 0, 1 or 2 values and is sorted.
-// TODO(phl): Use BoundedArray here.
 template<typename Argument, typename Value>
-std::vector<Argument> SolveQuadraticEquation(
+BoundedArray<Argument, 2> SolveQuadraticEquation(
     Argument const& origin,
     Value const& a0,
     Derivative<Value, Argument> const& a1,