--- a/numerics/root_finders_body.hpp
+++ b/numerics/root_finders_body.hpp
@@ -57,7 +57,7 @@ Argument Bisect(Function f,
 }
 
 template<typename Argument, typename Value>
-std::vector<Argument> SolveQuadraticEquation(
+BoundedArray<Argument, 2> SolveQuadraticEquation(
     Argument const& origin,
     Value const& a0,
     Derivative<Value, Argument> const& a1,
@@ -78,12 +78,12 @@ std::vector<Argument> SolveQuadraticEquation(
   if (discriminant.value == discriminant_zero &&
       discriminant.error == discriminant_zero) {
     // One solution.
-    return {origin - 0.5 * a1 / a2};
+    return {{origin - 0.5 * a1 / a2}, 1};
   } else if (discriminant.value < discriminant_zero ||
              (discriminant.value == discriminant_zero &&
               discriminant.error < discriminant_zero)) {
     // No solution.
-    return {};
+    return {{}, 0};
   } else {
     // Two solutions.  Compute the numerator of the larger one.
     Derivative1 numerator;
@@ -96,9 +96,9 @@ std::vector<Argument> SolveQuadraticEquation(
     auto const solution1 = origin + numerator / (2.0 * a2);
     auto const solution2 = origin + (2.0 * a0) / numerator;
     if (solution1 < solution2) {
-      return {solution1, solution2};
+      return {{solution1, solution2}, 2};
     } else {
-      return {solution2, solution1};
+      return {{solution2, solution1}, 2};
     }
   }
 }